requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudinary==1.15.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
coveralls==1.6.0
dj-database-url==0.5.0
Django==2.1.7
django-background-tasks==1.2.0
django-cloudinary-storage==0.2.3
django-compat==1.0.15
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==2.0.0
more-itertools==7.0.0
openapi-codec==1.3.2
pbr==5.1.3
Pillow==5.4.1
pip-tools==3.5.0
pluggy==0.9.0
psycopg2==2.7.7
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.0
pytest==4.3.1
pytest-cov==2.6.1
python-http-client==3.1.0
pytz==2019.1
requests==2.21.0
sendgrid==3.6.5
sendgrid-django==4.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
